Refactored the information component. It now only consists of a single component
