+++ src/script/navigation/page-selectors/ModelPageSelector.ts
@@ -1,6 +1,6 @@
 import PageSelectorStrategy from '../PageSelectorStrategy';
 import ContentViewNavigator from '../ContentViewNavigator';
-import Output from '../../../pages/Output.svelte';
+import Output from '../../../pages/ModelPage.svelte';
 import Menus, { MenuID } from '../Menus';
 
 class ModelPageSelector extends PageSelectorStrategy {
